Fix semantic bugs that silently produce incorrect results

+++ lpf/utils/validation.py
@@ -9,8 +9,8 @@ def is_param_invalid(params):
 
 
 def is_state_invalid(arr_u, arr_v):
-    abs_u = np.abs(arr_u.astype(np.float16))
-    abs_v = np.abs(arr_v.astype(np.float16))
+    abs_u = np.abs(arr_u.astype(np.float32))
+    abs_v = np.abs(arr_v.astype(np.float32))
     return (arr_u < 0).any() or (arr_v < 0).any() \
            or np.isnan(np.min(abs_u)) or np.isnan(np.min(abs_v)) \
            or np.isinf(np.max(abs_u)) or np.isinf(np.max(abs_v))
